+++ app/src/main/assets/index2.html
@@ -1,0 +1,81 @@
+//<!DOCTYPE html>
+//<html lang="en">
+//<head>
+//    <meta charset="UTF-8">
+//    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+//    <meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline' data:; media-src *;">
+//
+//    <title>Camera Page</title>
+//    <style>
+//        #video {
+//            width: 100%;
+//            height: auto;
+//            object-fit: cover;
+//        }
+//        button {
+//            font-size: 18px;
+//            padding: 10px 20px;
+//            margin: 5px;
+//        }
+//    </style>
+//</head>
+//<body>
+//
+//<button id="startButton">Start Camera</button>
+//<button id="takePicture" click=onPictureTaken2()>Take a picture</button>
+//<video id="video" autoplay playsinline></video>
+//    <canvas id="canvas" style="display:none;"></canvas>
+//<script>
+//    let video = document.querySelector('#video');
+//    let videoStream;
+//
+//    document.getElementById('startButton').addEventListener('click', function () {
+//      navigator.mediaDevices.getUserMedia({ video: { facingMode: 'environment' } })
+//        .then(function (stream) {
+//                videoStream = stream;
+//                video.srcObject = stream;
+//        })
+//        .catch(function (error) {
+//            console.log('Error accessing camera:', error.name);
+//            console.log('Error accessing camera:', error.message);
+//        });
+//    });
+//
+//
+//	function onPictureTaken2() {
+//	        // Draw the current video frame onto the canvas
+//            canvas.width = video.videoWidth;
+//            canvas.height = video.videoHeight;
+//            canvas.getContext('2d').drawImage(video, 0, 0, canvas.width, canvas.height);
+//
+//
+//
+//           var img = document.createElement('img');
+//           img.src = canvas.toDataURL();
+//           document.body.appendChild(img);
+//	}
+//
+//
+//	function onPictureTaken(image) {
+//    // Збережіть зображення в локальному сховищі
+//    var filename = 'image.jpg';
+//    var blob = new File([image], filename);
+//    window.localStorage.setItem(filename, blob);
+//
+//    // Відобразіть зображення в WebView
+//    var canvas = document.createElement('canvas');
+//    canvas.width = image.width;
+//    canvas.height = image.height;
+//
+//    var ctx = canvas.getContext('2d');
+//    ctx.drawImage(image, 0, 0);
+//
+//    var img = document.createElement('img');
+//    img.src = canvas.toDataURL();
+//    document.body.appendChild(img);
+//}
+//
+//</script>
+//
+//</body>
+//</html>
