+++ app/src/main/assets/styles.css
@@ -1,0 +1,20 @@
+body {
+  margin: 0;
+  font-family: 'Arial', sans-serif;
+}
+
+.sticky-header {
+  background-color: #333;
+  color: white;
+  padding: 10px;
+  text-align: center;
+  position: fixed;
+  top: 0;
+  width: 100%;
+  z-index: 1000;
+  transition: background-color 0.3s ease;
+}
+
+.sticky-header h1 {
+  margin: 0;
+}
