+++ app/src/main/res/layout/fragment_web_view_camera.xml
@@ -9,6 +9,7 @@
     <WebView
         android:id="@+id/webView"
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_height="match_parent"
+        android:hardwareAccelerated="true" />
 
 </FrameLayout>
